fix(submit-preprint): Minor fixes as a result of manual testing

+++ src/app/features/preprints/store/submit-preprint/submit-preprint.state.ts
@@ -300,7 +300,7 @@ export class SubmitPreprintState {
     const state = ctx.getState();
     const createdPreprintId = state.createdPreprint.data?.id;
     ctx.setState({ ...DefaultState });
-    if (createdPreprintId) {
+    if (createdPreprintId && !state.hasBeenSubmitted) {
       return this.preprintsService.deletePreprint(createdPreprintId);
     }
